fix: 홈 페이지 useSearchParams prerender 에러 해결 (#38)
