+++ src/app/(home)/page.tsx
@@ -1,51 +1,10 @@
-'use client';
-
-import {useState, Suspense} from 'react';
-import {useSearchParams, useRouter} from 'next/navigation';
-import {useRecruitmentStore} from '@/store/useRecruitmentStore';
-import {RecruitmentActive} from '@/app/(home)/_components/RecruitmentActive';
-import {RecruitmentInactive} from '@/app/(home)/_components/RecruitmentInactive';
-import {SubmissionCompleteModal} from '@/components/modal/SubmissionCompleteModal';
-import {SubmissionIncompleteModal} from '@/components/modal/SubmissionIncompleteModal';
+import HomeClient from '@/app/(home)/_components/HomeClient';
+import {Suspense} from 'react';
 
 export default function HomePage() {
-  const searchParams = useSearchParams();
-  const router = useRouter();
-
-  const isRecruiting = useRecruitmentStore((state) => state.isRecruiting);
-
-  const submittedParam = searchParams.get('submitted');
-
-  const [isSubmissionCompleteModalOpen, setIsSubmissionCompleteModalOpen] =
-    useState(submittedParam === 'true');
-  const [isSubmissionIncompleteModalOpen, setIsSubmissionIncompleteModalOpen] =
-    useState(submittedParam === 'false');
-
-  const closeSubmissionCompleteModal = () => {
-    setIsSubmissionCompleteModalOpen(false);
-    router.replace('/');
-  };
-
-  const closeSubmissionIncompleteModal = () => {
-    setIsSubmissionIncompleteModalOpen(false);
-    router.replace('/');
-  };
-
   return (
     <Suspense fallback={null}>
-      <main>
-        {isRecruiting ? <RecruitmentActive /> : <RecruitmentInactive />}
-      </main>
-      <SubmissionCompleteModal
-        isOpen={isSubmissionCompleteModalOpen}
-        onClose={closeSubmissionCompleteModal}
-        onConfirm={closeSubmissionCompleteModal}
-      />
-      <SubmissionIncompleteModal
-        isOpen={isSubmissionIncompleteModalOpen}
-        onClose={closeSubmissionIncompleteModal}
-        onConfirm={closeSubmissionIncompleteModal}
-      />
+      <HomeClient />
     </Suspense>
   );
 }
